+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Audio changelist: add search form field for 'Plattennummer'
 - include 'Genres' in Band and Musiker autocomplete dropdown table
 - send email to admins upon server errors
+- add feedback page: let users mail feedback to the admins
 
 ### Changed